requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pyodbc>=5.1.0
python-keycloak>=4.3.0
python-jose[cryptography]>=3.3.0
pydantic>=2.9.2
pydantic-settings>=2.6.1
python-multipart>=0.0.12
google-generativeai>=0.8.3
PyMuPDF>=1.24.12
httpx>=0.27.2
python-dotenv>=1.0.1
Pillow>=11.0.0
aiofiles>=24.1.0
